requirements.txt
Backtesting==0.3.0
matplotlib==3.3.2
numpy==1.19.2
pandas==1.1.2
PyYAML==5.3.1
seaborn==0.11.0
SQLAlchemy==1.3.19
